drawArena() implemented

+++ include/Atom.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct Atom{
-    double x;
-    double y;
-    double vx;
-    double vy;
+    float x;
+    float y;
+    float vx;
+    float vy;
 };
